Add new metadata key persistent_urls which removes the hash from all database urls when set to "true"
I understand the reason why the a substring of the hash is included in the url but
there are some use cases where the urls should persist across deployments. For bookmarks
for example or for scripts that use the JSON API.

This is the initial commit for this feature. Tests and documentation updates to follow.