Remove large, unmaintainable embedded structure
This is a huge pain to update when changing the inspect output (which, I
admit, is something only I seem to do). It's also tricky to see what has
caused a failure, even with pytest's diffing.

If we feel this is needed, we should probably test whether specific keys
exist, rather than the equality of the entire structure.