array facet: don't materialize unnecessary columns
The presence of `inner.*` causes SQLite to materialize a row with all
the columns. Those columns will be discarded later.

Instead, we can select only the column we'll use. This lets SQLite's
optimizer realize that the other columns in the CTE definition
aren't needed.

On a test table with 278K rows, 98K of which had an array, this speeds up
the facet calculation from 4 sec to 1 sec.