Group Issue #1910
Potential fix in datasette/views/database.py.

Issue #1910 refers to columns having no type checking and upon analysis it sets columns with no types as a "text" string. Implemented a feature that checks if the instance variables within that column are either Integers or Floats and sets them accordingly. If they are not it is then set as a "text" string instead as originally intended.

#1910 was not even an issue in the first place, I just classified Integers and Floats as themselves instead of everything being "text" strings.