Final fix for Group issue 2: 2123
My previous group issue 2 commit did not actually fix the issue, so I had to go back and fix it. Basically the issue wasn't calling the wrong argument in the datasette argument, but rather the command serve did not need to be called in the argument since the --reload command was already calling it. So, I reverted the changes I made in the previous commit and added the proper error message that better explains what is going on.

Now reload no longer calls serve and now the intended command "datasette  --reload <path/to/database>" works correctly